Move providers to separate module.
